requirements.txt
absl-py==1.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cmake==3.27.1
comm==0.1.4
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
Cython==3.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
einops==0.6.1
exceptiongroup==1.1.2
execnet==2.0.2
executing==1.2.0
expecttest==0.1.3
fastjsonschema==2.18.0
filelock==3.12.2
flash-attn==2.0.4
fonttools==4.42.0
fsspec==2023.6.0
gast==0.5.4
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.56.2
h5py==3.9.0
hypothesis==5.35.1
idna==3.4
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab==2.3.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupytext==1.15.0
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.2
llvmlite==0.40.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.1
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==5.1.0
mpmath==1.3.0
murmurhash==1.0.9
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==2.6.3
ninja==1.11.1
notebook==6.4.10
numpy==1.22.2
nvidia-dali-cuda120==1.28.0
nvidia-pyindex==1.0.9
oauthlib==3.2.2
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.2.0
polygraphy==0.47.1
pooch==1.7.0
preshed==3.0.8
prettytable==3.8.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
pyparsing==3.0.9
pytest==7.4.0
pytest-flakefinder==1.1.0
pytest-rerunfailures==12.0
pytest-shard==0.1.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytorch-quantization==2.1.2
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rpds-py==0.9.2
rsa==4.9
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sphinx-glpi-theme==0.3
srsly==2.4.7
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tbb==2021.10.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.1.0a0+29c30b1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
types-dataclasses==0.6.6
typing_extensions==4.7.1
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
xdoctest==1.0.2
